Moves staff lines to another file
